# Conflicts:
#	src/fesom_module.F90
#	src/gen_modules_read_NetCDF.F90
#	src/io_meandata.F90
#	src/oce_tracer_mod.F90
#	src/write_step_info.F90

+++ src/io_restart_file_group.F90
@@ -40,7 +40,7 @@ contains
     class(restart_file_group), target, intent(inout) :: this
     character(len=*), intent(in) :: name
     character(len=*), intent(in) :: units, longname
-    real(kind=8), target, intent(inout) :: local_data(:) ! todo: be able to set precision
+    real(kind=WP), target, intent(inout) :: local_data(:) ! todo: be able to set precision
     type(t_mesh), intent(in) :: mesh
     type(t_partit), intent(in) :: partit
     ! EO parameters
@@ -54,7 +54,7 @@ contains
     class(restart_file_group), intent(inout)           :: this
     character(len=*)         , intent(in)              :: name
     character(len=*)         , intent(in)              :: units, longname
-    real(kind=8)             , intent(inout), target   :: local_data(:,:) ! todo: be able to set precision
+    real(kind=WP)             , intent(inout), target   :: local_data(:,:) ! todo: be able to set precision
     type(t_mesh)             , intent(in)              :: mesh
     type(t_partit)           , intent(in)              :: partit
     integer                  , intent(in)   , optional :: ncat
@@ -78,7 +78,7 @@ contains
     class(restart_file_group), intent(inout) :: this
     character(len=*), intent(in) :: name
     character(len=*), intent(in) :: units, longname
-    real(kind=8), target, intent(inout) :: local_data(:) ! todo: be able to set precision
+    real(kind=WP), target, intent(inout) :: local_data(:) ! todo: be able to set precision
     type(t_mesh), intent(in) :: mesh
     type(t_partit), intent(in) :: partit
     ! EO parameters
@@ -93,7 +93,7 @@ contains
     class(restart_file_group), intent(inout) :: this
     character(len=*), intent(in) :: name
     character(len=*), intent(in) :: units, longname
-    real(kind=8), target, intent(inout) :: local_data(:,:) ! todo: be able to set precision
+    real(kind=WP), target, intent(inout) :: local_data(:,:) ! todo: be able to set precision
     type(t_mesh), intent(in) :: mesh
     type(t_partit), intent(in) :: partit
     ! EO parameters
@@ -146,7 +146,7 @@ contains
     class(restart_file_group), target, intent(inout) :: this
     character(len=*), intent(in) :: name
     character(len=*), intent(in) :: units, longname
-    real(kind=8), target, intent(inout) :: local_data(:) ! todo: be able to set precision
+    real(kind=WP), target, intent(inout) :: local_data(:) ! todo: be able to set precision
     type(t_mesh), intent(in) :: mesh
     type(t_partit), intent(in) :: partit
     ! EO parameters
@@ -160,7 +160,7 @@ contains
     class(restart_file_group), intent(inout)            :: this
     character(len=*)         , intent(in)               :: name
     character(len=*)         , intent(in)               :: units, longname
-    real(kind=8)             , intent(inout), target    :: local_data(:,:) ! todo: be able to set precision
+    real(kind=WP)             , intent(inout), target    :: local_data(:,:) ! todo: be able to set precision
     type(t_mesh)             , intent(in)               :: mesh
     type(t_partit)           , intent(in)               :: partit
     integer                  , intent(in)   , optional  :: ncat
@@ -188,7 +188,7 @@ contains
     class(restart_file_group), intent(inout) :: this
     character(len=*), intent(in) :: name
     character(len=*), intent(in) :: units, longname
-    real(kind=8), target, intent(inout) :: local_data(:) ! todo: be able to set precision
+    real(kind=WP), target, intent(inout) :: local_data(:) ! todo: be able to set precision
     type(t_mesh), intent(in) :: mesh
     type(t_partit), intent(in) :: partit
     ! EO parameters
@@ -203,7 +203,7 @@ contains
     class(restart_file_group), intent(inout) :: this
     character(len=*), intent(in) :: name
     character(len=*), intent(in) :: units, longname
-    real(kind=8), target, intent(inout) :: local_data(:,:) ! todo: be able to set precision
+    real(kind=WP), target, intent(inout) :: local_data(:,:) ! todo: be able to set precision
     type(t_mesh), intent(in) :: mesh
     type(t_partit), intent(in) :: partit
     ! EO parameters
